requirements.txt
Django>=3.2.11
django-model-utils==4.2.0
Pillow>=9.0.0
djangorestframework
